refactor(engine): extract logits processors + gpt-oss filter from inference.py (#454)
#3 maintainability: move the grammar logits-processor installer, the
OpenAI frequency/presence penalty processors, the model-vocab resolver,
and the gpt-oss channel stream filter into engine/logits_processors.py.
inference.py re-exports them so call sites and test imports are unchanged;
grammar-test patch targets repointed to the new module.
